Very hard bug: the _bridge.update{} was infinite looping.  I discovered (after decompiling to java) that this occurred because the thing that makes it thread-safe actually LOOPS until the start value of _bridge is stable.  This would never happen as it continually failed the compare condition.  Why? I changed some Sets to Lists in PhilipsHueBridgeInfo and classes downstream.  This made the equals not quite right anymore.
Solution?  I went ahead and commented out ALL the special cases of equality in these data classes.  As I understand data classes better, the custom equals function is not needed.  If I see bugs in the future I will uncomment the overridden code (which now works as I corrected the problem in the first paragraph).