Lots of refactoring.  I want to make it VERY clear which functions send data (eventually) to the bridge, and which functions are receiving data (as sse) from the bridge.
Added some serialization to simplify making json strings.  Turns out this caused a lot of frustration as the code would no longer compile; the serialization library is incompatible with kotlin 1.9.  So I had to switch to the latest version of kotlin (2.2.something), lots of changes in gradle and toml files.

Fixed a bug which prevented lights from being visible in Zones.  Weird: lights appear as DEVICEs in Rooms, but appear as LIGHTs in Zones.

Started creating light groups (actually zones) as part of getting Flocks to work.  Still in the early stages and still very buggy.